+++ pages/dashboard/create/cartoon.js
@@ -1,0 +1,1 @@
+// service to create a cartoon sketch of a face
